refactor: move skillToMarkdown to skillService; fix SKILL.md hosts frontmatter
Move skillToMarkdown from zipExportService.ts to skillService.ts where it
belongs (skill serialization is not a zip-export concern). Export it from
the skills barrel and update all three import sites: useOfficeChat.ts,
zipExportService.ts (now imports it), and zip-export-service.test.ts.

Fix the SKILL.md frontmatter: the Copilot SDK does not recognise 'hosts' as
a top-level skill attribute, producing a lint error since PR #38 made the
proxy pass skillDirectories to the SDK. Move 'hosts' under the SDK-supported
'metadata:' key. parseFrontmatter already handles this correctly because it
trims line indentation before parsing, so 'metadata:\n  hosts: [excel]' and
'hosts: [excel]' are parsed identically — no parser changes required.

Add tests:
- parseFrontmatter correctly reads hosts nested under metadata: block
- bundled Excel skill has hosts: ['excel'] after parsing
