+++ src/hooks/useOfficeChat.ts
@@ -5,14 +5,13 @@ import type { ThreadMessageLike, AppendMessage } from '@assistant-ui/react';
 import type { WebSocketCopilotClient, BrowserCopilotSession } from '@/lib/websocket-client';
 import { createWebSocketClient } from '@/lib/websocket-client';
 import { getToolsForHost } from '@/tools';
-import { getSkills, getImportedSkills } from '@/services/skills';
+import { getSkills, getImportedSkills, skillToMarkdown } from '@/services/skills';
 import { resolveActiveAgent } from '@/services/agents';
 import { resolveActiveMcpServers, toSdkMcpServers } from '@/services/mcp';
 import { useSettingsStore } from '@/stores';
 import { buildSystemPrompt } from '@/services/ai/systemPrompt';
 import { humanizeToolName } from '@/utils/humanizeToolName';
 import { inferProvider } from '@/types';
-import { skillToMarkdown } from '@/services/extensions/zipExportService';
 import type { AgentHost } from '@/types/agent';
 import type { OfficeHostApp } from '@/services/office/host';
 import { generateId } from '@/utils/id';
