+++ src/services/skills/index.ts
@@ -6,4 +6,5 @@ export {
   setImportedSkills,
   buildSkillContext,
   parseFrontmatter,
+  skillToMarkdown,
 } from './skillService';
